move extern calls to context
