+++ x/programs/wasmlanche/src/lib.rs
@@ -79,8 +79,8 @@ mod logging {
 
 pub use self::{
     context::{Context, ExternalCallContext},
-    program::{send, ExternalCallError, Program},
-    state::{get_balance, macro_types, Error},
+    program::{ExternalCallError, Program},
+    state::{macro_types, Error},
     types::{Address, Gas, Id, ProgramId, ID_LEN},
 };
 #[doc(hidden)]
